[tasks/features/T0003_features-ai-liz-operations_concept-execution_finish-abort-function]:  Implementing missing checking of safe concept deleting. | Implementing folder and file removal of relating concept. | Adding missing exception handlers.

+++ DominionsBackend/servers/ai/liz/server-ai-liz-operation/src/main/java/hu/kalmancheysandor/applications/dominions/servers/ai/liz/operation/service/neural/LizNeuralConceptService.java
@@ -10,17 +10,12 @@ import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.dto.
 import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.dto.concept.result.option.LizNeuralConceptResultHistoryScenarioOptionResponse;
 import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.entity.history.LizHistoryPlayer;
 import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.entity.history.LizHistoryScenario;
-import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.entity.training.neural.LizNeuralConcept;
-import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.entity.training.neural.LizNeuralExecution;
-import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.entity.training.neural.LizNeuralTrainingResult;
-import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.entity.training.neural.LizNeuralTrainingSnapshot;
+import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.entity.training.neural.*;
 import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.exception.concept.*;
 import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.repository.history.LizHistoryPlayerRepository;
 import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.repository.history.LizHistoryScenarioRepository;
 import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.repository.training.LizNeuralConceptRepository;
-import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.repository.training.neural.LizNeuralExecutionRepository;
-import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.repository.training.neural.LizNeuralTrainingResultRepository;
-import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.repository.training.neural.LizNeuralTrainingSnapshotRepository;
+import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.repository.training.neural.*;
 import hu.kalmancheysandor.applications.dominions.apis.server.ai.liz.shared.service.TLizService;
 import hu.kalmancheysandor.applications.dominions.apis.server.common.component.config.ApplicationConfig;
 import hu.kalmancheysandor.applications.dominions.apis.util.uuid.UUIDGenerator;
@@ -69,6 +64,12 @@ public class LizNeuralConceptService extends TLizService {
 
     @Autowired
     LizNeuralTrainingSnapshotRepository lizNeuralTrainingSnapshotRepository;
+
+    @Autowired
+    LizNeuralTrainingTaskRepository lizNeuralTrainingTaskRepository;
+
+    @Autowired
+    LizNeuralTrainingRepository lizNeuralTrainingRepository;
 
     @PersistenceContext
     private EntityManager entityManager;
@@ -162,7 +163,7 @@ public class LizNeuralConceptService extends TLizService {
         return modelMapper.map(recordModified, LizNeuralConceptUpdateResponse.class);
     }
 
-    public void deleteOneLizPersonnel(@NotBlank String recordUuid) {
+    public void deleteOneLizConcept(@NotBlank String recordUuid) {
 
         // Execution
         try {
@@ -198,12 +199,27 @@ public class LizNeuralConceptService extends TLizService {
             throw new LizNeuralConceptReferencedElsewhereException(recordId, recordToDelete.getName());
         }
 
-        // Delete item from db
+        // check whether an unfinished execution exists
+        if(lizNeuralExecutionRepository.isAnyRunning(recordId)) {
+            throw new LizNeuralConceptDeletionBlockedByActiveExecutionException(recordId);
+        }
+
+        // Delete child items of main
+        lizNeuralTrainingSnapshotRepository.deleteAllWhereConceptId(recordId);
+        lizNeuralTrainingResultRepository.deleteAllWhereConceptId(recordId);
+        lizNeuralTrainingTaskRepository.deleteAllWhereConceptId(recordId);
+        lizNeuralTrainingRepository.deleteAllWhereConceptId(recordId);
+        lizNeuralExecutionRepository.deleteAllWhereConceptId(recordId);
+
+        // Delete main item from db
         try {
             lizNeuralConceptRepository.deleteById(recordId);
         } catch (EntityNotFoundException e) {
             throw new LizNeuralConceptNotFoundException(recordId);
         }
+
+        // Delete file structure if exists. It has to be final due to any deleted files and folders are not recoverable.
+        deleteRecursivelyConceptDirectoryIfExists(recordId);
     }
